requirements.txt
Django==1.6.7
south>=1.0
urllib3==1.10.2
PyCrypto
PyOpenSSL
python-dateutil
oauth>=1.0.1
-e git+https://github.com/uw-it-aca/canvas-support-tools.git#egg=canvas_admin
-e git+https://github.com/uw-it-aca/library-guides-lti#egg=libguide
-e git+https://github.com/uw-it-aca/uw-groups-lti#egg=groups
-e git+https://github.com/uw-it-aca/grading-standard-lti#egg=grading_standard
-e git+https://github.com/uw-it-aca/grade-conversion-calculator#egg=grade_conversion_calculator
-e git+https://github.com/uw-it-aca/uw-restclients#egg=RestClients
-e git+https://github.com/uw-it-aca/django-supporttools#egg=django_supporttools
-e git+https://github.com/uw-it-aca/django-userservice#egg=Django-UserService
-e git+https://github.com/uw-it-aca/canvas-event-consumer#egg=canvas-event-consumer
-e git+https://github.com/uw-it-aca/canvas-admin-reconciler#egg=astra-reconciler
-e git+https://github.com/uw-it-aca/canvas-analytics#egg=canvas-analytics
-e git+https://github.com/uw-it-aca/authz_group#egg=AuthZ-Group
-e git+https://github.com/vegitron/django-saferecipient-email-backend.git#egg=safe_email_backend
-e git+https://github.com/abztrakt/django-mobileesp#egg=django_mobileesp
-e git+https://github.com/uw-it-aca/django-blti#egg=django_blti
-e git+https://github.com/uw-it-aca/django-aws-message#egg=django_aws_message
-e git+https://github.com/uw-it-aca/canvas-eos-import#egg=canvas-eos-import
